% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_fun.R
\name{eigen_decomp}
\alias{eigen_decomp}
\title{Title}
\usage{
eigen_decomp(
  A,
  K = min(dim(A)),
  symmetric = Matrix::isSymmetric(A),
  fast = TRUE
)
}
\arguments{
\item{fast}{}
}
\description{
Title
}
