% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_math_helpers.R, R/help-topics.R
\name{graph_laplacian}
\alias{graph_laplacian}
\title{Construct a graph Laplacian}
\usage{
graph_laplacian(A, normalized = TRUE, tau = 0)
}
\arguments{
\item{A}{A matrix or binary response vector.}

\item{normalized}{Whether to construct a normalized graph Laplacian.}

\item{tau}{Nonnegative graph regularization strength.}
}
\value{
A dense or sparse Laplacian matrix.
}
\description{
Constructs an unnormalized or normalized graph Laplacian from a dense or
sparse adjacency matrix.
}
\seealso{
\code{\link[=spectral_cluster]{spectral_cluster()}}, \code{\link[=ase]{ase()}}
}
