% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_math_helpers.R
\name{is_symmetric_matrix}
\alias{is_symmetric_matrix}
\title{Test matrix symmetry internally}
\usage{
is_symmetric_matrix(A)
}
\arguments{
\item{A}{A dense or sparse matrix.}
}
\value{
A logical scalar.
}
\description{
Tests base and Matrix-package matrix classes for numerical symmetry.
}
\keyword{internal}
