% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_math_helpers.R, R/help-topics.R
\name{modal}
\alias{modal}
\title{Compute the statistical mode}
\usage{
modal(x, na_rm = FALSE)
}
\arguments{
\item{x}{Input vector.}

\item{na_rm}{Remove missing comparisons.}
}
\value{
A scalar value with the same basic type as the input.
}
\description{
Returns the most frequent non-missing value in a vector.
}
\seealso{
\code{\link[=nmi]{nmi()}}
}
