% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help-topics.R, R/x0_helpers.R
\name{netcrop_param_select}
\alias{netcrop_param_select}
\title{Select NETCROP partition parameters}
\usage{
netcrop_param_select(test_prop = 0.02, n = NULL, o_range = c(0, 0.8))
}
\arguments{
\item{test_prop, o_range}{NETCROP automatic partition controls.}

\item{n}{Positive network size.}
}
\value{
A named parameter recommendation.
}
\description{
Recommends NETCROP partition parameters for a requested network size.
}
\seealso{
\code{\link[=netcrop_blockmodel]{netcrop_blockmodel()}}, \code{\link[=netcrop_rdpg]{netcrop_rdpg()}}, \code{\link[=netcrop_lsm]{netcrop_lsm()}}
}
