% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_math_helpers.R, R/help-topics.R
\name{outer_add}
\alias{outer_add}
\title{Add vectors by outer expansion}
\usage{
outer_add(x, y, operator = "+", use_cpp = TRUE)
}
\arguments{
\item{x, y}{Numeric vectors or objects operated on by a scalar loss.}

\item{operator}{Outer operation; currently \code{"+"}.}

\item{use_cpp}{Try the registered compiled accelerator.}
}
\value{
A numeric matrix.
}
\description{
Forms the matrix of all pairwise sums of two numeric vectors.
}
\seealso{
\code{\link[=procrustes]{procrustes()}}
}
