% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help-topics.R, R/x5_netcrop_regularizer.R
\name{pairwise_clustering_losses}
\alias{pairwise_clustering_losses}
\alias{pair_nmi_loss}
\alias{pair_hamming_loss}
\title{Compare pairs of clusterings}
\usage{
pair_nmi_loss(g_1, g_2, g_3, g_4)

pair_hamming_loss(g_1, g_2, g_3, g_4)
}
\arguments{
\item{g_1, g_2, g_3, g_4}{Community labels used by pairwise losses.}
}
\value{
A numeric loss.
}
\description{
Computes NMI- or Hamming-based disagreement between two pairs of labelings.
}
\seealso{
\code{\link[=nmi]{nmi()}}, \code{\link[=label_match_greedy]{label_match_greedy()}}
}
