% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_fun.R
\name{spectral.cluster}
\alias{spectral.cluster}
\title{Title}
\usage{
spectral.cluster(
  A,
  K,
  laplacian = F,
  tau = 0,
  row.norm = F,
  spherical = F,
  fast = T
)
}
\arguments{
\item{fast}{}
}
\description{
Title
}
