% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_math_helpers.R, R/help-topics.R
\name{is_symmetric_matrix}
\alias{is_symmetric_matrix}
\title{Test matrix symmetry}
\usage{
is_symmetric_matrix(A)
}
\arguments{
\item{A}{A matrix or binary response vector.}
}
\value{
A logical scalar.
}
\description{
Tests whether a dense or sparse matrix is symmetric within tolerance.
}
\seealso{
\code{\link[=is_symmetric_network_matrix]{is_symmetric_network_matrix()}}
}
