% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netOP-package.R
\name{matrix-generics}
\alias{matrix-generics}
\alias{sum}
\alias{mean}
\alias{diag}
\alias{rowMeans}
\alias{rowSums}
\alias{colMeans}
\alias{colSums}
\title{Matrix summary generics}
\usage{
sum(..., na.rm = FALSE)
}
\description{
These generics are re-exported from Matrix so that common summary
operations dispatch correctly for sparse network matrices after loading
netOP.
}
